pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "envdrift"
version = "0.1.0"
description = "Prevent environment variable drift with Pydantic schema validation, pre-commit hooks, and dotenvx encryption"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jainal Gosaliya", email = "gosaliya.jainal@gmail.com" }
]
keywords = [
    "environment",
    "variables",
    "pydantic",
    "validation",
    "dotenv",
    "config",
    "drift",
    "schema",
    "secrets",
    "pre-commit",
    "dotenvx",
    "encryption",
    "azure",
    "aws",
    "hashicorp",
    "vault",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "typer>=0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# Vault backends
azure = [
    "azure-identity>=1.15",
    "azure-keyvault-secrets>=4.8",
]
aws = [
    "boto3>=1.34",
]
hashicorp = [
    "hvac>=2.0",
]

# All vault backends combined
vault = [
    "envdrift[azure,aws,hashicorp]",
]

# Pre-commit hook installation (needs pyyaml)
precommit = [
    "pyyaml>=6.0",
]

# Documentation
docs = [
    "mkdocs-material>=9.5",
]

# Development dependencies
dev = [
    "ruff>=0.8.0",
    "pyrefly>=0.2.0",
    "bandit>=1.7.0",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "pre-commit>=3.0",
    "pyyaml>=6.0",
    "mkdocs-material>=9.5",
]

# All optional dependencies
all = [
    "envdrift[vault,precommit]",
]

[project.scripts]
envdrift = "envdrift.cli:app"

[project.urls]
Homepage = "https://github.com/jainal09/envdrift"
Documentation = "https://jainal09.github.io/envdrift"
Repository = "https://github.com/jainal09/envdrift"
Issues = "https://github.com/jainal09/envdrift/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/envdrift"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit (security)
]
ignore = [
    "S101",   # assert allowed in tests
    "E501",   # line too long - handled by formatter
    "RUF012", # mutable class attrs - acceptable in dataclasses
    "RUF022", # __all__ sorting - manual organization preferred
    "PTH123", # Path.open() - tomllib needs file handle
    "S110",   # try-except-pass - acceptable in config discovery
    "B904",   # raise from in typer exit - not applicable
    "S104",   # binding to 0.0.0.0 - test fixtures only
    "SIM102", # nested if - readability preference
    "SIM110", # any() vs loop - explicit loop is clearer here
    "S310",   # URL open audit - we control the URL source
    "S202",   # tarfile extractall - trusted source (GitHub releases)
    "S603",   # subprocess call - we control the input
    "RUF005", # list concat - readability preference
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-markers",
    "--cov=envdrift",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["src/envdrift"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]
